# open-api-docs.yaml
openapi: 3.0.3
info:
  title: Custom AI Assistant API
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
       - id
       - name
       - createdAt
paths:
...
  /projects:
    post:
      summary: Creates a new project.
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        201:
          description: The project has been successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        401:
          description: The user is not authorized to create a project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: Lists the projects of an user.
      operationId: listProjects
      responses:
        200:
          description: The projects of the user have been successfully listed.
          content:
            application/json:
              schema:
                title: ListProjectsResponse
                type: array
                items:
                  type: object
                  title: ProjectResponse
                  allOf:
                    - $ref: "#/components/schemas/ProjectResponse"
                    - type: object
                      properties:
                        chatbotName:
                          type: string
        401:
          description: The user is not authorized to list projects.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...